+++ blockapps-haskell/vault-wrapper/api/src/Strato/Strato23/API.hs
@@ -3,6 +3,7 @@
 module Strato.Strato23.API
   ( VaultWrapperAPI
   , module Strato.Strato23.API.Key
+  , module Strato.Strato23.API.Password
   , module Strato.Strato23.API.Ping
   , module Strato.Strato23.API.Signature
   , module Strato.Strato23.API.Types
@@ -10,6 +11,7 @@ module Strato.Strato23.API
 
 import           Servant
 import           Strato.Strato23.API.Key
+import           Strato.Strato23.API.Password
 import           Strato.Strato23.API.Ping
 import           Strato.Strato23.API.Signature
 import           Strato.Strato23.API.Types
@@ -18,4 +20,5 @@ type VaultWrapperAPI = GetPing
                   :<|> GetKey
                   :<|> PostKey
                   :<|> PostSignature
+                  :<|> PostPassword