+++ blockapps-haskell/vault-wrapper/server/src/Strato/Strato23/Database/Queries.hs
@@ -13,8 +13,10 @@ module Strato.Strato23.Database.Queries where
 
 import           BlockApps.Ethereum
 import           Control.Arrow
+import           Control.Monad                   (void)
 import qualified Crypto.Saltine.Class            as Saltine
 import qualified Crypto.Saltine.Core.SecretBox   as SecretBox
+import           Data.ByteString                 (ByteString)
 import qualified Data.ByteString.Char8           as C8
 import           Data.Int                        (Int32)
 import           Data.Maybe                      (fromMaybe, listToMaybe)
@@ -29,28 +31,51 @@ import           Strato.Strato23.Database.Tables
 
 getUserKeyQuery :: Text -> Query (Column PGBytea, Column PGBytea, Column PGBytea, Column PGBytea)
 getUserKeyQuery username = proc () -> do
-  (_, name, salt, nonce, encSecKey, address) <- queryTable usersTable -< ()
+  (_, name, salt, nonce, _, encSecPrvKey, address) <- queryTable usersTable -< ()
   restrict -< name .== constant username
-  returnA -< (salt, nonce, encSecKey, address)
+  returnA -< (salt, nonce, encSecPrvKey, address)
 
 postUserKeyQuery :: Text -> KeyStore -> Connection -> IO Bool
 postUserKeyQuery userName KeyStore{..} conn = do
   (userIds :: [Int32]) <- runQuery conn $ proc () -> do
-    (userId,name,_,_,_,_) <- queryTable usersTable -< ()
+    (userId,name,_,_,_,_,_) <- queryTable usersTable -< ()
     restrict -< name .== constant userName
     returnA -< userId
   case listToMaybe userIds of
     Just _ -> return False
     Nothing -> do
-      _ <- runInsertMany conn usersTable [
+      void $ runInsertMany conn usersTable [
         ( Nothing
         , constant userName
         , constant keystoreSalt
         , constant keystoreAcctNonce
         , constant keystoreAcctEncSecKey
+        , constant keystoreAcctEncSecKey
         , constant keystoreAcctAddress
         )]
       return True
+
+getMessageQuery :: Query (Column PGBytea, Column PGBytea, Column PGBytea)
+getMessageQuery = proc () -> do
+  (id', salt, nonce, enc_msg) <- queryTable messageTable -< ()
+  restrict -< id' .== constant (1 :: Int)
+  returnA -< (salt, nonce, enc_msg)
+
+postMessageQuery :: ByteString
+                 -> SecretBox.Nonce
+                 -> ByteString
+                 -> Connection
+                 -> IO Bool
+postMessageQuery salt nonce message conn = do
+  (msg :: [(ByteString, SecretBox.Nonce, ByteString)]) <- runQuery conn getMessageQuery
+  case msg of
+    (_:_) -> return False
+    [] -> True <$ runInsertMany conn messageTable [
+                    ( Nothing
+                    , constant salt
+                    , constant nonce
+                    , constant message
+                    )]
 
 instance QueryRunnerColumnDefault PGBytea Address where
   queryRunnerColumnDefault = queryRunnerColumn id