+++ blockapps-haskell/vault-wrapper/server/src/Strato/Strato23/Database/Tables.hs
@@ -22,6 +22,7 @@ usersTable :: Table
   , Column PGBytea
   , Column PGBytea
   , Column PGBytea
+  , Column PGBytea
   )
   ( Column PGInt4
   , Column PGText
@@ -29,12 +30,32 @@ usersTable :: Table
   , Column PGBytea
   , Column PGBytea
   , Column PGBytea
+  , Column PGBytea
   )
-usersTable = Table "users" $ p6
+usersTable = Table "users" $ p7
   ( optional "id"
   , required "x_user_unique_name"
   , required "salt"
   , required "nonce"
   , required "enc_sec_key"
+  , required "enc_sec_prv_key"
   , required "address"
+  )
+
+messageTable :: Table
+  ( Maybe (Column PGInt4)
+  , Column PGBytea
+  , Column PGBytea
+  , Column PGBytea
+  )
+  ( Column PGInt4
+  , Column PGBytea
+  , Column PGBytea
+  , Column PGBytea
+  )
+messageTable = Table "message" $ p4
+  ( optional "id"
+  , required "salt"
+  , required "nonce"
+  , required "enc_msg"
   )