+++ blockapps-haskell/vault-wrapper/server/src/Strato/Strato23/Server.hs
@@ -16,6 +16,7 @@ import           Servant.Swagger
 import           Strato.Strato23.API
 import           Strato.Strato23.Monad
 import           Strato.Strato23.Server.Key
+import           Strato.Strato23.Server.Password
 import           Strato.Strato23.Server.Ping
 import           Strato.Strato23.Server.Signature
 
@@ -24,6 +25,7 @@ vaultWrapper = getPing
           :<|> getKey
           :<|> postKey
           :<|> postSignature
+          :<|> postPassword
 
 serveVaultWrapper :: VaultWrapperEnv -> Server VaultWrapperAPI
 serveVaultWrapper env = hoistServer serverProxy (enterVaultWrapper env) vaultWrapper