+++ core-strato/blockapps-data/src/Blockchain/Data/Extra.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE TypeOperators    #-}
 
 module Blockchain.Data.Extra
     ( getGenesisHash